Split instrument(base) into two modules
Better named and avoids a huge module
